+++ tests/was-email-contracts/run-all.js
@@ -2,4 +2,5 @@
 "use strict";
 
 require("./was-uniqueness.test.js");
+require("./empty-week-policy.test.js");
 console.log("all was-email-contracts tests passed");